PAE-1412: extract log-capture test server helper
