+++ src/server/common/helpers/logging/request-logger.integration.test.js
@@ -1,28 +1,10 @@
-import hapi from '@hapi/hapi'
-import hapiPino from 'hapi-pino'
 import pino from 'pino'
 import { afterEach, describe, expect, it } from 'vitest'
 
 import { config } from '#config/config.js'
 import { loggerOptions } from './logger-options.js'
 import { createLogger } from './logger.js'
-
-const newServer = async () => {
-  const lines = []
-  const stream = { write: (s) => lines.push(s) }
-  const { transport: _transport, ...rest } = loggerOptions
-
-  const server = hapi.server({ port: 0 })
-  await server.register({
-    plugin: hapiPino,
-    options: { ...rest, enabled: true, level: 'trace', stream }
-  })
-
-  return { server, lines }
-}
-
-const findLine = (lines, predicate) =>
-  lines.map((s) => JSON.parse(s)).find(predicate)
+import { createLogCaptureServer } from '#server/common/test-helpers/log-capture-server.js'
 
 describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
   afterEach(() => {
@@ -31,7 +13,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
 
   it('should emit a canonical IndexedLogProperties payload through hapi-pino unchanged', async () => {
     config.set('cdpEnvironment', 'dev')
-    const { server, lines } = await newServer()
+    const { server, findLine } = await createLogCaptureServer()
     server.route({
       method: 'GET',
       path: '/test',
@@ -45,7 +27,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
     })
 
     await server.inject('/test')
-    const out = findLine(lines, (l) => l.message === 'test event')
+    const out = findLine((l) => l.message === 'test event')
 
     expect(out).toMatchObject({
       message: 'test event',
@@ -56,7 +38,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
 
   it('should set ECS service.name to the technical service identifier', async () => {
     config.set('cdpEnvironment', 'dev')
-    const { server, lines } = await newServer()
+    const { server, findLine } = await createLogCaptureServer()
     server.route({
       method: 'GET',
       path: '/test',
@@ -67,14 +49,14 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
     })
 
     await server.inject('/test')
-    const out = findLine(lines, (l) => l.message === 'svc-name-check')
+    const out = findLine((l) => l.message === 'svc-name-check')
 
     expect(out['service.name']).toBe('epr-frontend')
   })
 
   it('should bind ECS http.* and url.* on the per-request child logger (no flat req)', async () => {
     config.set('cdpEnvironment', 'dev')
-    const { server, lines } = await newServer()
+    const { server, findLine } = await createLogCaptureServer()
     server.route({
       method: 'GET',
       path: '/test',
@@ -85,7 +67,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
     })
 
     await server.inject('/test')
-    const out = findLine(lines, (l) => l.message === 'handler entered')
+    const out = findLine((l) => l.message === 'handler entered')
 
     expect(out).toMatchObject({
       http: { request: { method: 'GET' } },
@@ -97,7 +79,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
 
   it('should emit ECS http.response.* and url.path on the hapi-pino response auto-log', async () => {
     config.set('cdpEnvironment', 'dev')
-    const { server, lines } = await newServer()
+    const { server, findLine } = await createLogCaptureServer()
     server.route({
       method: 'GET',
       path: '/test',
@@ -105,10 +87,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
     })
 
     await server.inject('/test')
-    const out = findLine(
-      lines,
-      (l) => l.url?.path === '/test' && l.http?.response
-    )
+    const out = findLine((l) => l.url?.path === '/test' && l.http?.response)
 
     expect(out).toMatchObject({
       http: { response: { status_code: 200 } },
@@ -120,7 +99,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
 
   it('should map a logged err to ECS error.* when cdpEnvironment is non-prod', async () => {
     config.set('cdpEnvironment', 'dev')
-    const { server, lines } = await newServer()
+    const { server, findLine } = await createLogCaptureServer()
     server.route({
       method: 'GET',
       path: '/test',
@@ -131,7 +110,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
     })
 
     await server.inject('/test')
-    const out = findLine(lines, (l) => l.message === 'boom path')
+    const out = findLine((l) => l.message === 'boom path')
 
     expect(out?.error).toStrictEqual(
       expect.objectContaining({
@@ -144,7 +123,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
 
   it('should strip error.stack_trace when cdpEnvironment is prod', async () => {
     config.set('cdpEnvironment', 'prod')
-    const { server, lines } = await newServer()
+    const { server, findLine } = await createLogCaptureServer()
     server.route({
       method: 'GET',
       path: '/test',
@@ -155,7 +134,7 @@ describe('request-logger integration (hapi-pino + pino + ecs format)', () => {
     })
 
     await server.inject('/test')
-    const out = findLine(lines, (l) => l.message === 'boom path')
+    const out = findLine((l) => l.message === 'boom path')
 
     expect(out?.error).toBeDefined()
     expect(out?.error).not.toHaveProperty('stack_trace')
